pyproject.toml
[tool.poetry]
name = "checkeventsbot"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"
aiologger = "^0.7.0"
asyncpg = "^0.29.0"
telethon = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
